+++ examples/taskflow/composeApp/src/commonMain/kotlin/org/reduxkotlin/sample/taskflow/feature/undo/UndoMiddleware.kt
@@ -1,19 +1,13 @@
-package org.reduxkotlin.sample.taskflow.middleware
+package org.reduxkotlin.sample.taskflow.feature.undo
 
 import org.reduxkotlin.Middleware
 import org.reduxkotlin.Store
 import org.reduxkotlin.middleware
 import org.reduxkotlin.multimodel.ModelState
 import org.reduxkotlin.sample.taskflow.action.BoardRestored
-import org.reduxkotlin.sample.taskflow.action.PushUndo
-import org.reduxkotlin.sample.taskflow.action.Redo
-import org.reduxkotlin.sample.taskflow.action.SetUndoModel
-import org.reduxkotlin.sample.taskflow.action.Undo
 import org.reduxkotlin.sample.taskflow.core.Action
 import org.reduxkotlin.sample.taskflow.core.Undoable
 import org.reduxkotlin.sample.taskflow.model.BoardModel
-import org.reduxkotlin.sample.taskflow.model.UndoModel
-import org.reduxkotlin.sample.taskflow.reducer.undoReducer
 
 /**
  * The undo/redo middleware: the SINGLE place that maintains the per-account undo history.
