+++ examples/taskflow/composeApp/src/commonMain/kotlin/org/reduxkotlin/sample/taskflow/model/BoardSlice.kt
@@ -1,11 +1,8 @@
 package org.reduxkotlin.sample.taskflow.model
 
-import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.PersistentSet
-import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.persistentSetOf
 import org.reduxkotlin.sample.taskflow.core.AccountId
-import org.reduxkotlin.sample.taskflow.core.Board
 import org.reduxkotlin.sample.taskflow.core.CardId
 import org.reduxkotlin.sample.taskflow.core.LabelId
 import kotlin.time.Instant
@@ -16,11 +13,7 @@ data class FilterModel(
     val labelIds: PersistentSet<LabelId> = persistentSetOf(),
 )
 
-data class UndoModel(
-    val past: PersistentList<Board> = persistentListOf(),
-    val future: PersistentList<Board> = persistentListOf(),
-    val cap: Int = 15,
-)
+// UndoModel moved to …feature.undo.UndoModel
 
 data class SyncModel(
     // cards with an in-flight op (drives per-card optimistic alpha)
