+++ examples/taskflow/composeApp/src/commonTest/kotlin/org/reduxkotlin/sample/taskflow/reducer/BoardReducersTest.kt
@@ -9,15 +9,11 @@ import org.reduxkotlin.sample.taskflow.action.BoardRestored
 import org.reduxkotlin.sample.taskflow.action.BotAddedCard
 import org.reduxkotlin.sample.taskflow.action.BotMovedCard
 import org.reduxkotlin.sample.taskflow.action.LoadBoardSucceeded
-import org.reduxkotlin.sample.taskflow.action.PushUndo
-import org.reduxkotlin.sample.taskflow.action.Redo
 import org.reduxkotlin.sample.taskflow.action.Refresh
 import org.reduxkotlin.sample.taskflow.action.SetFilterAssignee
 import org.reduxkotlin.sample.taskflow.action.SetFilterQuery
-import org.reduxkotlin.sample.taskflow.action.SetUndoModel
 import org.reduxkotlin.sample.taskflow.action.SyncStatusChanged
 import org.reduxkotlin.sample.taskflow.action.ToggleFilterLabel
-import org.reduxkotlin.sample.taskflow.action.Undo
 import org.reduxkotlin.sample.taskflow.core.AccountId
 import org.reduxkotlin.sample.taskflow.core.ActivityEntry
 import org.reduxkotlin.sample.taskflow.core.AddCard
@@ -38,10 +34,17 @@ import org.reduxkotlin.sample.taskflow.core.OpId
 import org.reduxkotlin.sample.taskflow.feature.activity.ActivityModel
 import org.reduxkotlin.sample.taskflow.feature.activity.RecordActivity
 import org.reduxkotlin.sample.taskflow.feature.activity.activityReducer
+import org.reduxkotlin.sample.taskflow.feature.undo.PushUndo
+import org.reduxkotlin.sample.taskflow.feature.undo.Redo
+import org.reduxkotlin.sample.taskflow.feature.undo.SetUndoModel
+import org.reduxkotlin.sample.taskflow.feature.undo.Undo
+import org.reduxkotlin.sample.taskflow.feature.undo.UndoModel
+import org.reduxkotlin.sample.taskflow.feature.undo.pushUndo
+import org.reduxkotlin.sample.taskflow.feature.undo.undoModelReducer
+import org.reduxkotlin.sample.taskflow.feature.undo.undoReducer
 import org.reduxkotlin.sample.taskflow.model.BoardModel
 import org.reduxkotlin.sample.taskflow.model.FilterModel
 import org.reduxkotlin.sample.taskflow.model.SyncModel
-import org.reduxkotlin.sample.taskflow.model.UndoModel
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
